Addrating for the Detailpage (#45)
* add rate function for detailpage

* add rate function for detailpage (backend)

* finish rateing function for detail page

* finish rateing function for detail page (backend part)

* fixed Deepscan issues

+++ client/src/components/Card/Rate.js
@@ -26,7 +26,7 @@ export default function Rate({ rate }) {
   const balls = [];
 
   function Rating(rate) {
-    const intRate = parseInt(rate);
+    const intRate = parseFloat(rate);
     let count = 0;
     while (count < intRate) {
       balls.push(<YellowBall key={count} />);
